pyproject.toml
[project]
name = "areg"
version = "0.1.0"
description = "Event Director Analytics Demo - AI-powered insights for race directors"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.40.0",
    "pandas>=2.2.0",
    "plotly>=5.24.0",
    "folium>=0.18.0",
    "streamlit-folium>=0.23.0",
    "openai>=1.50.0",
    "boto3>=1.35.0",
    "numpy>=2.0.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools]
py-modules = []

[tool.setuptools.packages.find]
where = ["src"]
